+++ main/UseCases/MonthReportMaker.java
@@ -38,16 +38,16 @@ public class MonthReportMaker implements ReportMaker {
         for (Project proj : projList) {
             projectCompatibilityList.putAll(proj.getSkillsCompatibilities());
         }
-        ArrayList<HashMap<String, Integer>> internsSkills = getHiredInternsSkills(currentHRSystem.getHiredInternList());
+        ArrayList<HashMap<String, Double>> internsSkills = getHiredInternsSkills(currentHRSystem.getHiredInternList());
         return bakeProjectName(currentHRSystem.getProjectName(currentMonth)) + "\n" +
                 bakeProgress(projectProgress) + "\n" +
                 bakeInterns(internNames) + "\n" +
                 bakeInternsPerformances(internNames, internsSkills, projectCompatibilityList);
     }
 
-    private ArrayList<HashMap<String, Integer>> getHiredInternsSkills(ArrayList<HiredIntern> hiredInternList) {
+    private ArrayList<HashMap<String, Double>> getHiredInternsSkills(ArrayList<HiredIntern> hiredInternList) {
         //Makes an arrayList full of internSkills.
-        ArrayList<HashMap<String, Integer>> internCompatabilityList  = new ArrayList<>();
+        ArrayList<HashMap<String, Double>> internCompatabilityList  = new ArrayList<>();
         for (Intern i : hiredInternList){
             internCompatabilityList.add(i.getInternSkills());
         }
@@ -71,7 +71,7 @@ public class MonthReportMaker implements ReportMaker {
 
     @Override
     public String bakeInternsPerformances (String internNames,
-                                           ArrayList<HashMap<String, Integer>>  internSkills,
+                                           ArrayList<HashMap<String, Double>>  internSkills,
                                            HashMap<String, Integer> projectSkill) {
         StringBuilder returnLine = new StringBuilder(GamePrompts.INTERN_PERFORMANCE_HEADER + internNames + "\n");
         String[] internNamesList = internNames.split("|");
@@ -85,12 +85,12 @@ public class MonthReportMaker implements ReportMaker {
     }
 
     @Override
-    public int calculateInternPerformance(HashMap<String, Integer> internSkills,
+    public int calculateInternPerformance(HashMap<String, Double> internSkills,
                                           HashMap<String, Integer> projectSkill) {
         int result = 0;
         ArrayList<Double> effectiveSkills = new ArrayList<Double>();
         for (String key : internSkills.keySet()) {
-            int internSkill = internSkills.get(key);
+            double internSkill = internSkills.get(key);
             double compatibility = projectSkill.get(key);
             effectiveSkills.add(internSkill* compatibility);
         }
